Fix type conversion from int64 to int without bounds checking
Changed DirectRoomResponse.UnreadCount from int to int64 to match the return type from GetUnreadCountsForUser. This eliminates the unsafe type conversion that could overflow on 32-bit systems. Also updated the CLI DirectRoom struct to use int64 for consistency.

+++ cli/internal/api/client.go
@@ -71,7 +71,7 @@ type DirectRoom struct {
 	Type        string    `json:"type"`
 	OtherUser   User      `json:"other_user"`
 	LastMessage *Message  `json:"last_message,omitempty"`
-	UnreadCount int       `json:"unread_count"`
+	UnreadCount int64     `json:"unread_count"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
